countingRstToCountingMatrix() to utils
