+++ src/ffpack/utils/__init__.py
@@ -1,3 +1,4 @@
 from .generalUtils import *
-from .lccUtils import *
 from .fdrUtils import *
+from .lccUtils import *
+from .lsmUtils import *
